refac: createSHop 컴포넌트 imageUpload 컴포넌트화 작업
